Users/anush/Downloads/mental-health-risk-ai-main/mental-health-risk-ai-main/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyMongo==3.0.1

numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2

matplotlib==3.8.4

transformers==4.43.3
torch==2.3.1
huggingface_hub==0.24.5

pymongo==4.8.0

requests==2.32.3
tqdm==4.66.4
python-dotenv==1.0.1
